requirements.txt
# Enterprise Newsletter Generator - Dependencies

# Core AI/ML
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Document Processing
python-docx>=1.0.0
python-pptx>=0.6.0
PyPDF2>=3.0.0

# Web and API
requests>=2.31.0

# Data Processing
pandas>=2.0.0

# Optional but recommended
python-dotenv>=1.0.0
